% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/est_riemann_int.R
\name{approx_normmix}
\alias{approx_normmix}
\title{Approximate density of a normal mixture over 2d domain}
\usage{
approx_normmix(mix, win = spatstat::square(1))
}
\arguments{
\item{mix}{An object of class normmix}

\item{win}{An object of class owin, defaults to spatstat::square(1)}
}
\value{
A numerical vector corresponding to the density of each component within thw window.
}
\description{
Approximate the density of each component in a normal mixture within the domain using Riemann integral.
}
\examples{
mix1 <- normmix(ps=c(.5, .5),
                mus=list(c(0, 0), c(1, 1)),
                sigmas=list(.01*diag(2), .01*diag(2)))
approx_normmix(mix1)
}

