% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\docType{data}
\name{demo_mix}
\alias{demo_intsurf}
\alias{demo_mix}
\title{Demo objects}
\format{An object of class \code{normmix} of length 5.}
\usage{
demo_mix

demo_intsurf
}
\description{
Demo objects from the classes provided by this package.
}
\examples{
demo_mix <- normmix(ps = c(.3, .7), mus = list(c(0.2, 0.2), c(.8, .8)),
                    sigmas = list(.01*diag(2), .01*diag(2)))
demo_intsurf <- normmix(ps = c(.3, .7), mus = list(c(0.2, 0.2), c(.8, .8)),
                        sigmas = list(.01*diag(2), .01*diag(2)),
                        lambda = 100, win = square(1))
}
\keyword{datasets}

