% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funcs.R
\name{Demo_sppmix}
\alias{Demo_sppmix}
\title{Run an sppmix package demo or vignette}
\usage{
Demo_sppmix(whichdemo = NULL)
}
\arguments{
\item{whichdemo}{A number indicating which
demo to run. Do not pass any number in order
to see all the availabe choices.}
}
\description{
The function starts the different tutorials of
the \code{sppmix} package.

For examples see

\url{http://faculty.missouri.edu/~micheasa/sppmix/sppmix_all_examples.html
#Demo_sppmix}
}
\examples{
\donttest{
Demo_sppmix()#shows all available demos and opens the vignettes page
Demo_sppmix(1)#demo on sppmix objects}

}
\author{
Sakis Micheas
}
