% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots.R
\name{plotmix_2d}
\alias{plot.normmix}
\alias{plotmix_2d}
\title{2D contour plots for normal mixture and intensity surface}
\usage{
plotmix_2d(intsurf, pattern, contour = FALSE, truncate = TRUE, L = 256,
  ...)

\method{plot}{normmix}(mix, win, contour = FALSE, truncate = TRUE,
  L = 256)
}
\arguments{
\item{intsurf}{Object of class intensity_surface}

\item{pattern}{optional spatial point pattern to add to the plot. In the c
class of \code{\link[spatstat]{ppp}}.}

\item{contour}{Logical flag indicating whether to plot countour only.}

\item{truncate}{Whether to truncate the points outside the domain,
default to TRUE.}

\item{L}{number of grids on each coordinate.}

\item{...}{Further parameters to passed to \code{to_int_surf()}.}

\item{mix}{object of class \code{normmix}.}

\item{win}{Plotting window for normmix density. When omitted, it will
guess the range of the window by normal mixture parameters.}
}
\description{
Create a contour plot for the intensity surface with or without realizations
on it.
}
\examples{
# plot normmix density
mix1 <- normmix(ps = c(.3, .7),
                mus = list(c(0.2, 0.2), c(.8, .8)),
                sigmas = list(.01*diag(2), .01*diag(2)))
plot(mix1)

# plot intensity surface
intsurf1 <- normmix(ps = c(.3, .7),
                    mus = list(c(0.2, 0.2), c(.8, .8)),
                    sigmas = list(.01*diag(2), .01*diag(2)),
                    lambda = 100,
                    win = square(1))
plotmix_2d(intsurf1)

pp1 <- rsppmix(intsurf = intsurf1)
plotmix_2d(intsurf1, pp1)  # with points

}

