% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post.R
\name{drop_realization}
\alias{drop_realization}
\title{Drop a subset of MCMC realization}
\usage{
drop_realization(fit, drop)
}
\arguments{
\item{fit}{Object of class \code{damcmc_res} or \code{bdmcmc_res}.}

\item{drop}{If one integer is provided, it will drop the first 1:drop
realizations. If an integer vector is provided, it will drop these
iterations. If a logical vector is provided (with the same length as chain
length of \code{fit}), it will be used for subsetting directly.}
}
\description{
Drop realizations from a MCMC chain.
}
\examples{
fit <- est_mix_bdmcmc(redwood, m = 5)
fit
drop_realization(fit, 500)
drop_realization(fit, fit$numcompo != 5)

}

