% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_intensity_np.R
\name{est_intensity_np}
\alias{est_intensity_np}
\title{Estimate intensity surface by using non-parametric method.}
\usage{
est_intensity_np(pattern, win, h, L = 10, kernel = c("Epanechnikov"),
  edgecorrect = TRUE, truncate = TRUE)
}
\arguments{
\item{pattern}{An object of class \code{\link[spatstat]{ppp}}. It should be a
2 dimesional spatial point pattern.}

\item{win}{Object of class \code{\link[spatstat]{owin}}.}

\item{h}{Kernel bandwith.  \code{h} should be a postive number.}

\item{L}{Number of grid for x and y axis.}

\item{kernel}{Kernel used to estimate intensity surface.  Currently, only
support Epanechnikov kernel.}

\item{edgecorrect}{Logical flag indicating whether to use edge-correction in
estimating intensity surface. The default is TRUE.}

\item{truncate}{Logical flag indicating whether truncation is used for
\code{pattern}. The default is TRUE.}
}
\value{
An object of class \code{\link[spatstat]{im}}.
}
\description{
Using Epanechnikov kernel to estimate intensity surface.
}
\examples{
if (require(spatstat)){
  mix1 <- rnormmix(3, .01, 5, square(5))
  pattern1 <- rsppmix(30, mix1, square(5))
}

# estimate and plot the estimated intensity surface
if (require(spatstat)){
  surf1 <- est_intensity_np(pattern1, win=square(5), h=0.05, L=100)
  plot(surf1)
}
}

