% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mix_intensity.R
\name{est_mix_damcmc}
\alias{est_mix_bdmcmc}
\alias{est_mix_damcmc}
\title{Estimate mixture model using MCMC}
\usage{
est_mix_damcmc(pp, m, truncate = FALSE, L = 5000, LL = 100)

est_mix_bdmcmc(pp, m, truncate = FALSE, lambda, lambdab, hyper, L = 5000,
  LL = 100)
}
\arguments{
\item{pp}{point pattern object of class \code{ppp}.}

\item{m}{either number of components to fit in data augmentation MCMC or
maximum number of component in Birth Death MCMC.}

\item{truncate}{logical, indicating whether truncation is used, where the
component density are restricted within the domain of the point pattern.}

\item{L}{length of MCMC chain, default to 5000.}

\item{LL}{additional fitting parameter}

\item{lambda}{parameter for BDMCMC}

\item{lambdab}{parameter for BDMCMC}

\item{hyper}{parameter for BDMCMC}
}
\description{
These functions fits a mixture model to spatial point pattern data using
DAMCMC or BDMCMC.
}
\examples{
# see vignett "Work with 2D Normal Mixtures" for details.

}

