% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/est_riemann_int.R
\name{est_riemann_int}
\alias{est_riemann_int}
\title{Estimating Riemann integral over a 2d doamin.}
\usage{
est_riemann_int(func, xlimits, ylimits, L = 100)
}
\arguments{
\item{func}{A vectorized function with two inputs. It should take two vector of the same length and return a numerical vector.}

\item{xlimits}{A vector with two values defining the range of the domain on x-axis.}

\item{ylimits}{A vector with two values defining the range of the domain on y-axis.}

\item{L}{A numerical value controls how many sub-regions are partitioned on each axis, default to 100.}
}
\value{
The estimated value of Riemann integral of the function on that domain.
}
\description{
Estimating Riemann integral of a function over a 2d doamin. The domain is partitioned into small regions and the function is evaluated on the center of each region.
}
\examples{
func <- function(x, y) exp(-.5 * x^2 - .5 * y^2) / (2 * pi)
est_riemann_int(func, c(-3, 3), c(-3, 3))
}

