% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixLabelswitching.R
\name{FixLS_da}
\alias{FixLS_da}
\title{Fix label switching}
\usage{
FixLS_da(fit, burnin = length(fit$allgens_List)/10, xlab1 = "x",
  ylab1 = "y", plot_result = FALSE)
}
\arguments{
\item{fit}{An object contains all posterior realizations from
\code{\link{est_mix_damcmc}} or \code{\link{est_mix_bdmcmc}}.}

\item{burnin}{Length of burnin, default value is 1/10 of total number of
iteration.}

\item{xlab1}{The label for x-axis}

\item{ylab1}{The label for y-axis}

\item{plot_result}{Logical flag indicating whether plot the point pattern
and intensity surface after permutation. The default is FALSE.}
}
\description{
Fix label switching by apply the best permutation to the posterior
realization (MCMC result)
}
\examples{
# generate data
mix2 <- normmix(ps=c(.4, .6), mus=list(c(0.3, 0.3), c(0.7, 0.7)),
sigmas = list(.02*diag(2), .01*diag(2)))
pp2 <- rsppmix(100,mix2,square(1))
# Run Data augmentation MCMC and get posterior realizations
post = est_mix_damcmc(pp2,L = 5000,2,truncate = F)
# get posterior mean for each parameter
post_mean = get_post(post)
# plot the estimated intensity surface
plot(post_mean$post_normmix, post_mean$mean_lambda, square(1))
# Fix label switching
post_fixed = FixLS_da(post,square(1))
# get posterior mean for each parameter after permutation
post_mean2 = get_post(post_fixed)
# plot the estimated intensity surface after permutating
plot(post_mean2$post_normmix, post_mean2$mean_lambda, square(1))
}
\author{
Athanasios Christou Micheas, Jiaxun Chen, Yuchen Wang
}

