% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post.R
\name{get_post}
\alias{get_post}
\alias{get_post.bdmcmc_res}
\alias{get_post.damcmc_res}
\title{Get Posterior normal mixture.}
\usage{
get_post(obj, ...)

\method{get_post}{damcmc_res}(fit, burnin = fit$L/10)

\method{get_post}{bdmcmc_res}(fit, num_comp, burnin = fit$L/10)
}
\arguments{
\item{fit}{Object of class \code{damcmc_res} or \code{bdmcmc_res}.}

\item{burnin}{Number of burn-in iterations. By default, it's 1/10 of chain
length.}

\item{num_comp}{Number of components. The posterior will be calculated only
based on iterations have that many number of components.}
}
\value{
An object of class \code{intensity_surface}.
}
\description{
Calculate posterior normal mixture of a DAMCMC or BDMCMC fit. The posterior
mixture is calculated by using posterior mean of parameters. For birth-death
MCMC, the number of component should be specified, and all realization with
that number of components are gathered to calculate posterior mixture.
}
\examples{
fit <- est_mix_damcmc(pp = redwood, m = 3)
post_intsurf <- get_post(fit, burnin = 1000)

fit <- est_mix_bdmcmc(pp = redwood, m = 5)
post_intsurf <- get_post(fit, num_comp = 4, burnin = 1000)
}

