% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funcs.R
\name{plotstring}
\alias{plotstring}
\title{General Helper Functions}
\usage{
plotstring(str = "Hello World")
}
\arguments{
\item{str}{A string to display.}
}
\description{
The \code{plotstring} function plots a string in a generic plot device.
}
\examples{
\donttest{
plotstring()}

}
