% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R, R/est_riemann_int.R
\name{inormmix}
\alias{inormmix}
\title{Calculate intensity of locations under a mixture.}
\usage{
inormmix(int_surf, win, x, y, L = 128, truncate = TRUE)

inormmix(int_surf, win, x, y, L = 128, truncate = TRUE)
}
\arguments{
\item{win}{An object of class \code{\link[spatstat]{owin}}.}

\item{x}{Locations where the density function will be evaluated. It will
automatically expand it into a grid. If missing, a regular grid over the
window will be given.}

\item{y}{Locations where the density function will be evaluated. It will
automatically expand it into a grid. If missing, a regular grid over the
window will be given.}

\item{L}{Number of grid on x and y axes.}

\item{truncate}{Whether the density is truncated in the window (truncate)
or not.}

\item{lambda}{Mean intensity value}

\item{lambda}{Mean intensity value}
}
\value{
A pixel image of class \code{\link[spatstat]{im}} with the values
 corresponding to the intensity at given locations.

A pixel image of class \code{\link[spatstat]{im}} with the values
 corresponding to the intensity at given locations.
}
\description{
Calculate intensity values of a given mixture at given locations.

Calculate intensity values of a given mixture at given locations.
}
\examples{
if (require(spatstat)) {
  mix1 <- rnormmix(8, sig0 = .01, 10, square(2))
  int <- inormmix(100, mix1, square(2))
 plot(int)
}
if (require(spatstat)) {
  mix1 <- rnormmix(8, sig0 = .01, 10, square(2))
  int <- inormmix(100, mix1, square(2))
 plot(int)
}
}

