% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\name{normmix}
\alias{normmix}
\alias{summary.intensity_surface}
\alias{summary.normmix}
\title{Two-dimensional mixture with normal components.}
\usage{
normmix(ps, mus, sigmas, lambda = NULL, win = NULL, estimated = FALSE)

\method{summary}{normmix}(mix)

\method{summary}{intensity_surface}(mix)
}
\arguments{
\item{ps}{Vector of component probabilities.}

\item{mus}{A list where every element is a vector of length 2, defining the
center of each component.}

\item{sigmas}{A list where every element is a 2 by 2 covariance matrix,
defining the covariance for each component.}

\item{lambda}{Optional parameter of average number of points. If set with
window, the returned object will be an intensity surface.}

\item{win}{Optional parameter of window. Must be set with intensity value to
create an intensity surface.}

\item{estimated}{Whether it's an estimated mixture. By default it's set to
FALSE. When using an estimated mixture, it should be set to TRUE.}

\item{mix}{An object of class \code{normmix}}
}
\value{
An object of class "normmix" containing the following components:
 \item{m}{Number of components.}
 \item{ps}{Vector of component probabilities.}
 \item{mus}{List of mean vectors of components.}
 \item{sigmas}{List of covariance matrix of components.}
 \item{intensity}{Optional intensity value if lambda is provided when calling
 \code{normmix}.}
 \item{window}{Optional window object of class \code{\link[spatstat]{owin}}.}
 \item{estimated}{Whether the normal mixture is estimated.}
}
\description{
Constructor function of \code{normmix} class. Create a mixture in
two-dimensional with bivariate normal components. If additional parameters
lambda and window are set, it will create an intensity surface.
}
\examples{
mix1 <- normmix(ps = c(.3, .7),
                mus = list(c(0.2, 0.2), c(.8, .8)),
                sigmas = list(.01*diag(2), .01*diag(2)))
mix1


intsurf1 <- normmix(ps = c(.3, .7),
                    mus = list(c(0.2, 0.2), c(.8, .8)),
                    sigmas = list(.01*diag(2), .01*diag(2)),
                    lambda = 100,
                    win = square(1))
intsurf1

}
\seealso{
\code{\link{rnormmix}} for generating random mixture.
}

