% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funcs.R
\name{openwin_sppmix}
\alias{openwin_sppmix}
\title{Opens a new graphics window}
\usage{
openwin_sppmix(check2open = FALSE)
}
\arguments{
\item{check2open}{Logical: TRUE to open a newplot, FALSE do not open.}
}
\description{
This function is independent of the OS present, and is useful
when working outside of RStudio. The latter GUI
places all plots under the plots tab, but if working
in the R GUI the plots will be overwritten if you don't open a new device.

For examples see

\url{http://faculty.missouri.edu/~micheasa/sppmix/sppmix_all_examples.html
#openwin_sppmix}
}
\details{
This function is used by almost all plotting functions of the \code{\link{sppmix}} package.
}
\examples{
\donttest{
openwin_sppmix(TRUE)}

}
\author{
Sakis Micheas
}
