% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{plot.bdmcmc_res}
\alias{plot.bdmcmc_res}
\title{Plot the average surfaces from BDMCMC fit}
\usage{
\method{plot}{bdmcmc_res}(fit, win = fit$data$window,
  burnin = length(fit$allgens)/10, LL = 100, zlims = c(0, 0))
}
\arguments{
\item{fit}{Object of class \code{damcmc_res} or \code{bdmcmc_res}.}

\item{burnin}{Number of burn-in iterations. By default, it's 1/10 of chain
length.}

\item{LL}{Number of grid on x and y axes.}

\item{zlims}{The limits of z axis. The default does not have
additional limits on z axis.}
}
\description{
Plot the average surfaces from BDMCMC fit
}
\examples{
fit <- est_mix_bdmcmc(pp = redwood, m = 5)
plot(fit)
}

