% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{plot.damcmc_res}
\alias{plot.damcmc_res}
\title{Plot result from DAMCMC fit}
\usage{
\method{plot}{damcmc_res}(fit, burnin = length(fit$allgens)/10)
}
\arguments{
\item{fit}{Object of class \code{damcmc_res} or \code{bdmcmc_res}.}

\item{burnin}{Number of burn-in iterations. By default, it's 1/10 of chain
length.}
}
\description{
Plot result from DAMCMC fit
}
\examples{
fit <- est_mix_damcmc(pp = redwood, m = 3)
plot(fit)
}

