% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normmix.R
\name{plot.normmix}
\alias{plot.normmix}
\title{Plot mixture density.}
\usage{
\method{plot}{normmix}(mix, win, L = 100, truncate = TRUE, ...)
}
\arguments{
\item{mix}{An object of class \code{\link{normmix}}.}

\item{win}{An object of class \code{\link[spatstat]{owin}}.}

\item{L}{Length of grid on x and y axes.}

\item{truncate}{Whether the density is truncated in the window (truncate)
or not.}

\item{...}{Further arguments passed to \code{\link[rgl]{persp3d}}.}
}
\description{
Plot the density surface of a mixture.
}
\examples{
if (require(spatstat)) {
  mix1 <- rnormmix(8, .01, 10, square(2), rand_m = T)
  plot(mix1, square(2))
}
}

