% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots.R
\name{plot.sppmix}
\alias{plot.sppmix}
\title{Plot spatial point pattern generated from mixture.}
\usage{
\method{plot}{sppmix}(pattern, intsurf, lambda, ...)
}
\arguments{
\item{pattern}{A point pattern of class sppmix or \code{\link[spatstat]{ppp}}.}

\item{intsurf}{Object of class \code{\link{intensity_surface}}.}

\item{...}{Parameters passed to \code{\link[ggplot2]{ggplot}}.}
}
\description{
Plot funciton for spatial point pattern generated from mixture. It's an
alternative to spatstat's plotting functions.
}
\examples{
intsurf1 <- normmix(ps = c(.3, .7),
                    mus = list(c(0.2, 0.2), c(.8, .8)),
                    sigmas = list(.01*diag(2), .01*diag(2)),
                    lambda = 100,
                    win = square(1))

spp <- rsppmix(intsurf1)
plot(spp)
plot(spp, intsurf1)

}

