% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_avg_surf.R
\name{plot_avgsurf}
\alias{plot_avgsurf}
\title{Plot the average intensity surface}
\usage{
plot_avgsurf(fit, win, LL = 30, burnin = length(fit$allgens_List)/10)
}
\arguments{
\item{fit}{An object contains all posterior realizations from
\code{\link{est_mix_damcmc}} or \code{\link{est_mix_bdmcmc}}.}

\item{win}{An object of class \code{\link[spatstat]{owin}}.}

\item{LL}{Number of grid on x and y axes.}

\item{burnin}{Length of burnin, default value is 1/10 of total number of
iteration.}
}
\description{
Plot the average of the posterior realizations for the intensity surface
}
\examples{
# generate data
mix2 <- normmix(ps=c(.4, .6), mus=list(c(0.1, 0.1), c(0.8, 0.8)),
sigmas=list(.02*diag(2), .01*diag(2)))
pp2 <- rsppmix(100,mix2,square(1))
# Run Data augmentation MCMC and get posterior realizations
post=est_mix_damcmc(pp2,L = 5000,2,truncate = F)
# Plot the average of realized surfaces
plot_avgsurf(fit = post, win = square(1), LL = 30, burnin = 1000)
}
\author{
Athanasios Christou Micheas, Jiaxun Chen, Yuchen Wang
}

