% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{plot_chains}
\alias{plot_chains}
\title{Plot MCMC chains for mixture mean and probability.}
\usage{
plot_chains(fit, burnin = fit$L/10, chain = c("p", "x", "y"),
  ncol = fit$m\%\%3 + 1)
}
\arguments{
\item{fit}{Object of class \code{damcmc_res} or \code{bdmcmc_res}.}

\item{burnin}{Number of burn-in iterations. By default, it's 1/10 of chain
length.}

\item{chain}{Character vector choosing from c("p", "x", "y"). Multiple
choices is supported. It will plot MCMC chain for corresponding variables.}

\item{ncol}{Number of columns in each plot.}
}
\description{
Plot MCMC chains generated by \code{\link{est_mix_damcmc}} or
\code{\link{est_mix_bdmcmc}}.
}
\examples{
fit <- est_mix_damcmc(pp = redwood, m = 3)
plot_chains(fit)

# Only plot X and Y
plot_chains(fit, chain = c("x", "y"))
}

