% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ind.R
\name{plot_ind}
\alias{plot_ind}
\title{Plot membership indicator}
\usage{
plot_ind(dares)
}
\arguments{
\item{dares}{list of results from DAMCMC}
}
\description{
Plot membership indicator
}
\examples{
mix1 <- rnormmix(3, .01, 4, square(1))
pp1=rsppmix(200, mix1, spatstat::square(1))
data=cbind(pp1$x,pp1$y)
post=DAMCMC2d_sppmix(data,c(0,1),c(0,1),3, 5000, 1000, 50, trunc = FALSE)
plot_ind(post)
}

