% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots_3d.R
\name{plot_mix_3d}
\alias{plot_mix_3d}
\title{3D density plots for normal mixture and intensity surface.}
\usage{
plot_mix_3d(x, ..., truncate = TRUE, L = 128,
  main = "Normal Mixture 3D Surface Plot")
}
\arguments{
\item{x}{Object of class \code{intensity_surface} or \code{normmix}.}

\item{...}{Further parameters passed to \code{to_int_surf()}.
See \code{\link{to_int_surf}} for details.}

\item{truncate}{Whether to truncate the points outside the domain,
default to TRUE.}

\item{L}{Number of grids on each coordinate.}

\item{main}{Title for the density plot.}
}
\description{
Plot the density surface of a normal mixture in an interactive 3D plot.
}
\examples{
s <- plot_mix_3d(demo_intsurf)
s

}

