% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots.R
\name{plotmix_2d}
\alias{plotmix_2d}
\title{2D plot for intensity surface}
\usage{
plotmix_2d(intsurf, pattern, filled = TRUE, truncate = TRUE, L = 256,
  density = FALSE, xlim, ylim, ...)
}
\arguments{
\item{intsurf}{Object of class intensity_surface}

\item{pattern}{optional spatial point pattern to add to the plot. In the c
class of \code{\link[spatstat]{ppp}}.}

\item{filled}{Logical flag indicating whether plot filled contour plot.
The default is TRUE.}

\item{truncate}{Whether to truncate the points outside the domain,
default to TRUE.}

\item{L}{number of grids on each coordinate.}

\item{density}{whether to plot mixture density instead of intensity.}

\item{xlim, ylim}{omitted except when density = TRUE}

\item{...}{Further parameters passed to \code{to_int_surf()}.}
}
\description{
Create a contour plot for the intensity surface with or without realizations
on it.
}
\examples{
intsurf1 <- normmix(ps = c(.3, .7),
                    mus = list(c(0.2, 0.2), c(.8, .8)),
                    sigmas = list(.01*diag(2), .01*diag(2)),
                    lambda = 100,
                    win = square(1))
pp1 <- rsppmix(intsurf = intsurf1)

# plot the theoretical intensity surface with a realization
plotmix_2d(intsurf1, pp1)
}

