% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{Plots_off}
\alias{Plots_off}
\title{Closes all open plots}
\usage{
Plots_off()
}
\description{
The function closes all Rgl plots, as well as, any graphics devices.

For examples see

\url{http://www.stat.missouri.edu/~amicheas/sppmix/sppmix_all_examples.html
#Plots_off}
}
\examples{

\dontrun{
Plots_off()}

}
\author{
Sakis Micheas
}
\seealso{
\code{\link{Save_AllOpenRglGraphs}}
}

