% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_num_generator.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Dirichlet random number generator.}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{Number of generator}

\item{alpha}{Parameter for Dirichlet distribution}
}
\description{
Generate random numbers that follow Dirichlet distribution.
}

