% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{import}
\name{reexports}
\alias{owin}
\alias{reexports}
\alias{square}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{spatstat}{\code{\link[spatstat]{owin}}, \code{\link[spatstat]{square}}}
}}
\keyword{internal}

