% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normmix.R
\name{rnormmix}
\alias{rnormmix}
\title{Generate mixture with normal components.}
\usage{
rnormmix(m, sig0, sigdf, win = spatstat::square(1), rand_m = FALSE)
}
\arguments{
\item{m}{Number of component if \code{rand_m = FALSE}. When \code{rand_m = TRUE}, number of component is random and this is the maximum number of components.}

\item{sig0}{Tunning parameter in generating random matrix from Wishart distribution.}

\item{sigdf}{Degree of freedom in generating random matrix from Wishart distribution.}

\item{win}{Object of class \code{spatstat::owin}. The mean vectors are inside this window.}

\item{rand_m}{Whether the number of components are random or fixed (default).}
}
\value{
Object of class \code{normmix}.
}
\description{
Generate a mixture on a 2d window where the mean and variance of the components are random. The number of component can either be fixed or random.
}
\examples{
mix1 <- rnormmix(3, .01, 5)
mix2 <- rnormmix(3, .01, 5, rand_m = TRUE)
}

