% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{Save_AllOpenRglGraphs}
\alias{Save_AllOpenRglGraphs}
\title{Saves RGL plots}
\usage{
Save_AllOpenRglGraphs(dir1, filename1 = "RglGraph")
}
\arguments{
\item{dir1}{Directory to save the plots.}

\item{filename1}{Template filename. Each open plot is saved as filename1.png, filename2.png, and so forth.}
}
\description{
The function saves all open RGL plots (3d plots) to the specified directory
and using a template name.

For examples see

\url{http://faculty.missouri.edu/~micheasa/sppmix/sppmix_all_examples.html
#Save_AllOpenRglGraphs}
}
\examples{
\donttest{
# use a temporary directory to save the plots
Save_AllOpenRglGraphs(dir1=tempdir())}

}
\seealso{
\code{\link{Plots_off}}
}
\author{
Sakis Micheas
}
