% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_mix_model.R
\name{selectMix}
\alias{selectMix}
\title{Mixture Model Selection}
\usage{
selectMix(pp, Ms, L = 10000, burnin = 1000, truncate = FALSE,
  runallperms = 0)
}
\arguments{
\item{runallperms}{Set to 0 to use an approximation to the
Likelihood and Entropy within the MCMC (not affected by label switching).
Set to 1 to use an identifiability constraint
in order to remove label switching from the posterior
means. Set to 2 to run through all permutations and use decision
theory (minimize squared error loss) to obtain
the best permutation and then compute the posterior
means (BEST approach, but very slow operation). The default is 0.}
}
\value{
A list containing the following components:
\item{AIC}{the values of the AIC criterion}
\item{BIC}{the values of the BIC criterion}
\item{ICLC}{the values of the ICLC criterion}
\item{BayesFactor}{a matrix with all the Bayes Factors}
\item{Marginal}{the values of the marginal density}
\item{LogLikelihood}{the values of the LogLikelihood}
}
\description{
Finds the best number of components by computing
model selection criteria, including
AIC (Akaike Information Criterion),
BIC (Bayesian Information Criterion),
ICLC (Integrated Classification Likelihood Criterion),
and the Bayes factors of models with different
number of components against each other.
The entertained model is based on the posterior means
of MCMC runs, for several numbers of components
defined in the vector Ms.
}
\examples{
# create the true mixture
truemix <- normmix(ps=c(.2, .6,.2), mus=list(c(0.3, 0.3), c(0.7, 0.7), c(0.5, 0.5)),
sigmas = list(.005*diag(2), .001*diag(2), .001*diag(2)),
lambda=100,win=square(1))
# generate the point pattern
genPPP <- rsppmix(truemix)
plot(genPPP)
# compute model selection criteria
ModelSel=selectMix(genPPP,1:5,runallperms=0)
# show info
ModelSel
ModelSel=selectMix(genPPP,1:5,runallperms=1)
# show info
ModelSel
ModelSel=selectMix(genPPP,1:5,runallperms=2)
# show info
ModelSel
}
\author{
Sakis Micheas, Jiaxun Chen, Yuchen Wang
}
\references{
McLachlan, G., and Peel, D. (2000). Finite Mixture Models. Wiley-Interscience.

Jasra, A., Holmes, C.C. and Stephens, D. A. (2005). Markov Chain Monte Carlo Methods and the Label Switching Problem in Bayesian Mixture. Statistical Science, 20, 50-67.
}

