% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_post.R
\name{summary.damcmc_res}
\alias{summary.damcmc_res}
\title{Summary of an MCMC result}
\usage{
\method{summary}{damcmc_res}(fit, dgt = 2)
}
\arguments{
\item{fit}{An object contains all posterior realizations from
\code{\link{est_mix_damcmc}} or \code{\link{est_mix_bdmcmc}}.}
}
\description{
Prints a brief summary of an MCMC realization result
}
\examples{
# generate data
mix2 <- normmix(ps=c(.4, .6), mus=list(c(0.1, 0.1), c(0.8, 0.8)),
sigmas=list(.02*diag(2), .01*diag(2)))
pp2 <- rsppmix(100,mix2,square(1))
# Run Data augmentation MCMC and get posterior realizations
post=est_mix_damcmc(pp2,L = 5000,2,truncate = F)
summary the posterior results
summary(post)
}
\author{
Athanasios Christou Micheas, Jiaxun Chen, Yuchen Wang
}

