% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_switching.R
\name{test_labswitch}
\alias{test_labswitch}
\title{Test if posterior realizations of mu have label switching}
\usage{
test_labswitch(fit, burnin = fit$L/10)
}
\arguments{
\item{fit}{An object contains all posterior realizations from
\code{\link{est_mix_damcmc}} or \code{\link{est_mix_bdmcmc}}.}

\item{burnin}{Length of burnin, default value is 1/10 of total number of
iteration.}
}
\description{
Test if there is a label switching of the posterior realizations of mu by
testing if the mean of each component changed dramatically during MCMC.
}
\author{
Athanasios (Sakis) Micheas
}

