% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R, R/est_riemann_int.R
\name{approx_normmix}
\alias{approx_normmix}
\title{Approximate density of a normal mixture over a 2d domain.}
\usage{
approx_normmix(mix, win)

approx_normmix(mix, win)
}
\arguments{
\item{mix}{An object of class \code{\link{normmix}}}

\item{win}{An object of class \code{\link[spatstat]{owin}}}

\item{mix}{An object of class \code{\link{normmix}}}

\item{win}{An object of class \code{\link[spatstat]{owin}}}
}
\value{
A numerical vector corresponding to the density of each component
 within the window.

A numerical vector corresponding to the density of each component
 within the window.
}
\description{
Approximate the density of each component in a normal mixture within the
domain using multivariate normal density function.

Approximate the density of each component in a normal mixture within the
domain using multivariate normal density function.
}
\examples{
if (require("spatstat")) {
  mix1 <- normmix(ps=c(.3, .7),
                  mus=list(c(0, 0), c(1, 1)),
                 sigmas=list(.01*diag(2), .01*diag(2)))
  win1 <- spatstat::square(1)
  plot(dnormmix(mix1, win1))
  approx_normmix(mix1, win1)
}
if (require("spatstat")) {
  mix1 <- normmix(ps=c(.3, .7),
                  mus=list(c(0, 0), c(1, 1)),
                 sigmas=list(.01*diag(2), .01*diag(2)))
  win1 <- spatstat::square(1)
  plot(dnormmix(mix1, win1))
  approx_normmix(mix1, win1)
}
}

