% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{approx_normmix}
\alias{approx_normmix}
\title{Approximate density of a normal mixture over a 2d domain.}
\usage{
approx_normmix(mix, xlim = c(0, 1), ylim = c(0, 1))
}
\arguments{
\item{mix}{An object of class \code{\link{normmix}}}

\item{xlim, ylim}{vector of length two. Mixture density are estimated within
this range.}
}
\value{
A numerical vector corresponding to the density of each component
 within the window.
}
\description{
Approximate the density of each component in a normal mixture within the
domain using multivariate normal density function.
}

