% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots_2d.R
\name{plot.intensity_surface}
\alias{plot.intensity_surface}
\alias{plot.normmix}
\title{2D density plots for normal mixture and intensity surface}
\usage{
\method{plot}{intensity_surface}(x, ..., pattern, contour = FALSE,
  truncate = TRUE, L = 256)

\method{plot}{normmix}(x, xlim, ylim, contour = FALSE, truncate = FALSE,
  L = 256, ...)
}
\arguments{
\item{x}{Object of class \code{intensity_surface} or \code{normmix}.}

\item{...}{Further parameters passed to \code{to_int_surf()}.
See \code{\link{to_int_surf}} for details.}

\item{pattern}{Optional spatial point pattern to add to the plot. In the
class of \code{\link[spatstat]{ppp}}.}

\item{contour}{Logical flag indicating whether to plot countour only.}

\item{truncate}{Whether to truncate the points outside the domain,
default to TRUE.}

\item{L}{Number of grids on each coordinate.}

\item{xlim, ylim}{Limits for the plot when \code{x} is a \code{normmix}. If
missing, quantile estimation is used to find a reasonable region.}
}
\description{
Create a contour plot for the intensity surface with or without realizations
on it.
}
\examples{
# plot normmix density
plot(demo_mix)

# plot intensity surface
plot(demo_intsurf)

pp1 <- rsppmix(demo_intsurf)
plot(demo_intsurf, pattern = pp1)  # with points

}

