% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots_2d.R
\name{plot.intensity_surface}
\alias{plot.intensity_surface}
\alias{plot.normmix}
\title{2D density plots for normal mixture and intensity surface}
\usage{
\method{plot}{intensity_surface}(intsurf, pattern, contour = FALSE,
  truncate = TRUE, L = 256, ...)

\method{plot}{normmix}(mix, xlim, ylim, contour = FALSE, truncate = FALSE,
  L = 256)
}
\arguments{
\item{intsurf}{Object of class \code{intensity_surface} or \code{normmix}.}

\item{pattern}{Optional spatial point pattern to add to the plot. In the
class of \code{\link[spatstat]{ppp}}.}

\item{contour}{Logical flag indicating whether to plot countour only.}

\item{truncate}{Whether to truncate the points outside the domain,
default to TRUE.}

\item{L}{Number of grids on each coordinate.}

\item{...}{Further parameters passed to \code{to_int_surf()}.}

\item{mix}{object of class \code{normmix}.}
}
\description{
Create a contour plot for the intensity surface with or without realizations
on it.
}
\examples{
# plot normmix density
plot(demo_mix)

# plot intensity surface
plotmix_2d(demo_intsurf)

pp1 <- rsppmix(intsurf = demo_intsurf)
plotmix_2d(demo_intsurf, pp1)  # with points

}

