% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnormmix}
\alias{dnormmix}
\title{Calculate density of normal mixture.}
\usage{
dnormmix(mix, xlim = as.numeric(c(0, 1)), ylim = as.numeric(c(0, 1)),
  L = 128L, truncate = TRUE)
}
\arguments{
\item{mix}{An object of class \code{normmix} or \code{intensity_surface}.
When an intensity surface is given, it estimates the intensity instead of
density within the domain.}

\item{xlim, ylim}{The range within which the density or intensity is
calculated.}

\item{L}{Length of grid on each axis. The density is calculated on a L * L
grid.}

\item{truncate}{Whether to truncate the density to be within the domain.}
}
\value{
An matrix with the first two columns the grid coordinates and the
third column the density (intensity) value.
}
\description{
When a \code{normmix} object is given, this function estimates density
values within the domain. When a \code{intensity_surface} is given, it
multiplies the estimated density values with intensity and returns the
estimated intensity.
}
\examples{
est_density <- dnormmix(demo_mix)
est_density <- dnormmix(demo_intsurf)
}

