% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R
\name{dnormmix}
\alias{dnormmix}
\title{Calculate density of locations under a mixture.}
\usage{
dnormmix(mix, xlim = c(0, 1), ylim = c(0, 1), L = 128, truncate = TRUE)
}
\arguments{
\item{mix}{An object of class \code{\link{normmix}}}

\item{xlim}{vector of length two. Mixture density are estimated within
this range.}

\item{ylim}{vector of length two. Mixture density are estimated within
this range.}
}
\value{
An object of class \code{\link[spatstat]{im}}. This is a pixel image
 on a grid with values corresponding to the density at that location.
}
\description{
Calculate the density values of a given mixture with normal components.
}
\examples{
est_density <- dnormmix(demo_mix)
}
\seealso{
\code{\link[spatstat]{summary.im}} and
 \code{\link[spatstat]{plot.im}} for manipulating with pixel image object.
}

