% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximations.R, R/est_riemann_int.R
\name{dnormmix}
\alias{dnormmix}
\title{Calculate density of locations under a mixture.}
\usage{
dnormmix(mix, win, x, y, L = 128, truncate = TRUE)

dnormmix(mix, win, x, y, L = 128, truncate = TRUE)
}
\arguments{
\item{mix}{An object of class \code{\link{normmix}}.}

\item{win}{An object of class \code{\link[spatstat]{owin}}.}

\item{x, y}{Locations where the density function will be evaluated. It will
automatically expand it into a grid. If missing, a regular grid over the
window will be given.}

\item{L}{Number of grid on x and y axes.}

\item{truncate}{Whether the density is truncated in the window (truncate)
or not.}

\item{x, y}{Locations where the density function will be evaluated. It will
automatically expand it into a grid. If missing, a regular grid over the
window will be given.}
}
\value{
An object of class \code{\link[spatstat]{im}}. This is a pixel image
 on a grid with values corresponding to the density at that location.

An object of class \code{\link[spatstat]{im}}. This is a pixel image
 on a grid with values corresponding to the density at that location.
}
\description{
Calculate the density values of a given mixture with normal components.

Calculate the density values of a given mixture with normal components.
}
\examples{
if (require(spatstat)) {
  mix1 <- rnormmix(8, sig0 = .01, 10, square(2))
  den <- dnormmix(mix1, square(2))
 plot(den)
}
if (require(spatstat)) {
  mix1 <- rnormmix(8, sig0 = .01, 10, square(2))
  den <- dnormmix(mix1, square(2))
 plot(den)
}
}
\seealso{
\code{\link[spatstat]{summary.im}} and
 \code{\link[spatstat]{plot.im}} for manipulating with pixel image object.

\code{\link[spatstat]{summary.im}} and
 \code{\link[spatstat]{plot.im}} for manipulating with pixel image object.
}

