% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mix_intensity.R
\name{est_mix_damcmc}
\alias{est_mix_bdmcmc}
\alias{est_mix_damcmc}
\title{Estimate mixture model using MCMC}
\usage{
est_mix_damcmc(pp, m, truncate = FALSE, L = 5000)

est_mix_bdmcmc(pp, m, truncate = FALSE, lambda = 1, lambdab = 10,
  hyper = c(5, 0.01, 3, 2, 1, 1), L = 5000, LL = 100)
}
\arguments{
\item{pp}{point pattern object of class \code{ppp}.}

\item{m}{either number of components to fit in data augmentation MCMC or
maximum number of component in Birth Death MCMC.}

\item{truncate}{logical, indicating whether truncation is used, where the
component density are restricted within the domain of the point pattern.}

\item{L}{length of MCMC chain, default to 5000.}

\item{lambda}{Parameter in Poisson prior, lambda = 1 by default.}

\item{lambdab}{Birth rate, lambdab = 10 by default.}

\item{hyper}{hyperparameters for the hierarchical prior. See 'Details'.}
}
\description{
These functions fits a mixture model to spatial point pattern data using
DAMCMC or BDMCMC.
}
\details{
Birth-Death MCMC uses the same notations as the paper from Stephens, M.(2000).
The definition of hyperparameters can be found in Stephens's paper formula (21)-(24).
}
\examples{
# see vignett "Work with 2D Normal Mixtures" for details.

}
\references{
Stephens, M. "BAYESIAN ANALYSIS OF MIXTURE MODELS WITH AN UNKNOWN NUMBER OF
COMPONENTS—AN ALTERNATIVE TO REVERSIBLE JUMP METHODS", The Annals of Statistics 2000,
Vol. 28, No. 1, 40–74
}

