% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mix_intensity.R
\name{est_mix_damcmc}
\alias{est_mix_bdmcmc}
\alias{est_mix_damcmc}
\title{Estimate mixture model using MCMC}
\usage{
est_mix_damcmc(pp, m, truncate = FALSE, L = 5000, hyper_da = c(3, 1, 1))

est_mix_bdmcmc(pp, m, truncate = FALSE, lambda1 = 1, lambda2 = 10,
  hyper = c(5, 0.01, 3, 2, 1, 1), L = 5000)
}
\arguments{
\item{pp}{point pattern object of class \code{ppp}.}

\item{m}{either number of components to fit in data augmentation MCMC or
maximum number of component in Birth Death MCMC.}

\item{truncate}{logical, indicating whether truncation is used, where the
component density are restricted within the domain of the point pattern.}

\item{L}{length of MCMC chain, default to 5000.}

\item{hyper_da}{hyperparameters for DAMCMC, default is (3, 1, 1).}

\item{lambda1}{Parameter in Poisson prior, lambda = 1 by default.}

\item{lambda2}{Birth rate, lambdab = 10 by default.}

\item{hyper_da}{hyperparameters for the hierarchical prior. See 'Details'.}
}
\value{
A list of MCMC realizations
\item{allgens_List}{A list of generated nornal mixture.}
\item{genps}{Matrix of probabilities of components in the normal
mixture.}
\item{genmus}{Matrix of means of components in the normal
mixture.}
\item{gensigmas}{Matrix of covariance matrices of components in the normal
 mixture.}
\item{genzs}{Matrix of component indicators of points.}
\item{genlambdas}{Matrix of lambda(average intensity).}
\item{data}{The original point pattern.}
}
\description{
These functions fits a mixture model to spatial point pattern data using
DAMCMC or BDMCMC.
}
\details{
Birth-Death MCMC uses the same notations as the paper from Stephens, M.(2000).
The definition of hyperparameters can be found in Stephens's paper formula (21)-(24).
}
\examples{
# see vignett "Work with 2D Normal Mixtures" for details.

}
\references{
Stephens, M. "Bayesian analysis of mixture models with an unknown
number of components—an alternative to reversible jump methods", The Annals of Statistics 2000,
Vol. 28, No. 1, 40–74
}

