% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post.R
\name{get_post}
\alias{get_post}
\alias{get_post.bdmcmc_res}
\alias{get_post.damcmc_res}
\title{Get Posterior normal mixture}
\usage{
get_post(obj, ...)

\method{get_post}{damcmc_res}(fit, burnin = fit$L/10)

\method{get_post}{bdmcmc_res}(fit, comp, burnin = fit$L/10)
}
\arguments{
\item{fit}{object from \code{est_mix_damcmc} or \code{est_mix_bdmcmc}.}

\item{burnin}{number of burn-in iterations. By default, it's 1/10 of chain
length.}

\item{comp}{number of components. The posterior will be calculated only
based on iterations where}
}
\value{
An object of class \code{intensity_surface}.
}
\description{
Calculate posterior mixture after DAMCMC or BDMCMC fit. The posterior
mixture is calculated by using posterior mean of parameters. For birth-death
MCMC, the number of component need to be specified.
}
\examples{

fit <- est_mix_damcmc(pp = redwood, m = 3)
post_intsurf <- get_post(fit, burnin = 1000)
fit <- est_mix_bdmcmc(pp = redwood, m = 5)
post_intsurf <- get_post(fit, 3, burnin = 1000)
}

