% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mix_intensity.R
\name{get_post}
\alias{get_post}
\title{Calculate posterior mixture from estimated mixture}
\usage{
get_post(obj, ...)
}
\arguments{
\item{fit}{object from \code{est_mix_damcmc} or \code{est_mix_bdmcmc}.}

\item{burnin}{number of burn-in iterations.}
}
\value{
An object of class \code{intensity_surface}.
}
\description{
Calculate posterior mixture from estimated mixture
}
\details{
For Birth-Death MCMC, the number of component need to be specified.
}
\examples{

fit <- sppmix::est_mix_damcmc(pp = redwood, m = 3, truncate = FALSE,
                              L = 50000, LL = 100)
get_post(fit, burnin = 5000)

}

