% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kstest2d.R
\name{kstest2d}
\alias{kstest2d}
\title{Two-sample Two-dimesional Kolmogorov-Smirnov test}
\usage{
kstest2d(x1, x2)
}
\arguments{
\item{x1,x2}{Objects of class \code{\link[spatstat]{ppp}}.}
}
\value{
A list with class "htest" containing the following components:
\item{statistic}{the value of the KS statistic}
\item{p.value}{the p-value of the test}
\item{alternative}{a character string describing the alternative hypothesis}
}
\description{
Performs a test of goodness-of-fit of two point patterns.  The emperical
distribution of two point patterns are compared using the Kolmogorov-Smirnov
test.
}
\examples{
# genrating two point patterns
if(require(spatstat)){
  mix1 <- rnormmix(3, .01, 5, square(5))
  mix2 <- rnormmix(8, .01, 10, square(5))
  pp1=rsppmix(20,mix1,square(5))
  pp2=rsppmix(20,mix2,square(5))
}
# Test for goodness of fit
kstest2d(pp1,pp2)
}
\references{
J. A. Peacock, "Two-dimensional goodness-of-fit testingin
astronomy", Monthly Notices Royal Astronomy Society 202 (1983)615-627.

Adapted from Matlab code by Dylan Muir.
}

