% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{plot.bdmcmc_res}
\alias{plot.bdmcmc_res}
\title{Plot the average surfaces from BDMCMC fit}
\usage{
\method{plot}{bdmcmc_res}(fit, win = fit$data$window,
  burnin = length(fit$allgens)/10, LL = 100, zlims = c(0, 0))
}
\arguments{
\item{fit}{An object contains all posterior realizations from
\code{\link{est_mix_bdmcmc}}}

\item{burnin}{Length of burnin, default value is 1/10 of total number of
iteration.}

\item{LL}{Number of grid on x and y axes.}

\item{zlims}{The limits of z axis. The default does not have
additional limits on z axis.}
}
\description{
Plot the average surfaces from BDMCMC fit
}
\examples{

fit <- sppmix::est_mix_bdmcmc(pp = redwood, m = 3, truncate = FALSE,
                              L = 5000)
plot(fit)
}

