% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.damcmc_res}
\alias{plot.damcmc_res}
\title{Plot result from DAMCMC fit}
\usage{
\method{plot}{damcmc_res}(fit, burnin = length(fit$allgens)/10)
}
\description{
Plot result from DAMCMC fit
}
\examples{
fit <- sppmix::est_mix_damcmc(pp = redwood, m = 3, truncate = FALSE,
                              L = 5000, burnin = 100, LL = 100)
plot(fit)
}

