% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots.R
\name{plot.intensity_surface}
\alias{plot.intensity_surface}
\title{Plot normal mixture density.}
\usage{
\method{plot}{intensity_surface}(intsurf, truncate = TRUE, L = 256,
  zlims = c(0, 0), pos = c(0, 0, 0),
  main = "Intensity Surface of Normal Mixture", grayscale = FALSE, ...)
}
\arguments{
\item{intsurf}{Object of class \code{intensity_surface} or \code{normmix}.}

\item{truncate}{Whether to truncate the points outside the domain,
default to TRUE.}

\item{L}{Number of grid on x and y axes.}

\item{zlims}{The limits of z axis. The default does not has additional limits
on z axis.}

\item{pos}{The point coodinates for the title. The default is the top middle
in the 3-D plot.}

\item{main}{Title for the density plot.}

\item{grayscale}{Logical flag indicate if using gray scale as the color
of the 3-D plot.}

\item{...}{Further parameters passed to \code{to_int_surf()}.}
}
\description{
Plot the density surface of a normal mixture.
}
\examples{
plot(demo_intsurf)
}

