% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots.R
\name{plot.sppmix}
\alias{plot.sppmix}
\title{Plot spatial point pattern generated from mixture.}
\usage{
\method{plot}{sppmix}(pattern, mus, ...)
}
\arguments{
\item{pattern}{A point pattern of class sppmix or
\code{\link[spatstat]{ppp}}.}

\item{mus}{An optioanl list of theoretical mean for each component.}

\item{...}{Additional parameters to \code{add_title} to be plot on title}
}
\description{
Plot funciton for spatial point pattern generated from mixture. It's an
alternative to spatstat's plotting functions.
}
\examples{
intsurf1 <- normmix(ps = c(.3, .7),
                    mus = list(c(0.2, 0.2), c(.8, .8)),
                    sigmas = list(.01*diag(2), .01*diag(2)),
                    lambda = 100,
                    win = square(1))

spp <- rsppmix(intsurf1)
plot(spp)
plot(spp, mus = intsurf1$mus)
plot(spp, lambda = 100)

}

