% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots.R
\name{plot.sppmix}
\alias{plot.sppmix}
\title{Plot spatial point pattern generated from mixture.}
\usage{
\method{plot}{sppmix}(pattern, intsurf, lambda, ...)
}
\arguments{
\item{pattern}{A point pattern of class sppmix or \code{\link[spatstat]{ppp}}.}

\item{intsurf}{Object of class \code{\link{intensity_surface}}.}

\item{...}{Parameters passed to \code{\link[ggplot2]{ggplot}}.}
}
\description{
Plot funciton for spatial point pattern generated from mixture. It's an
alternative to spatstat's plotting functions.
}
\examples{
# plot point patterns generated by sppmix
mix1 <- rnormmix(3, .01, 5, square(1))
spp <- rsppmix(mix1, 100, square(1))
plot(spp, mix1)

# plot point patterns generated by spatstat
spp <- spatstat::rpoispp(100)
plot.spp(spp)
}

