% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix_plots_2d.R
\name{plot.sppmix}
\alias{plot.sppmix}
\title{Plot spatial point pattern generated from mixture.}
\usage{
\method{plot}{sppmix}(pattern, mus, ...)
}
\arguments{
\item{pattern}{A point pattern of class sppmix or
\code{\link[spatstat]{ppp}}.}

\item{mus}{An optioanl list of theoretical mean for each component.}

\item{...}{Additional parameters to \code{add_title} to be plot on title.}
}
\description{
Plot funciton for spatial point pattern generated from mixture. It's an
alternative to spatstat's plotting functions.
}
\examples{
spp <- rsppmix(demo_intsurf)
plot(spp)
plot(spp, mus = intsurf1$mus)
plot(spp, mus = intsurf1$mus, lambda = 100)

}

