% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sppmix}
\alias{plot.sppmix}
\title{Plot spatial point pattern generated from mixture.}
\usage{
\method{plot}{sppmix}(spp, ...)
}
\arguments{
\item{spp}{A point pattern of class sppmix or \code{\link[spatstat]{ppp}}.}

\item{...}{Parameters passed to \code{\link[ggplot2]{qplot}}.}
}
\description{
Plot funciton for spatial point pattern generated from mixture. It's an
alternative to spatstat's plotting functions.
}
\examples{
# plot point patterns generated by sppmix
mix1 <- mix1 <- rnormmix(3, .01, 5, square(1))
spp <- rsppmix(100, mix1, square(1))
plot(spp)

# plot point patterns generated by spatstat
spp <- spatstat::rpoispp(100)
plot.spp(spp)
}

