% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{plot_chains}
\alias{plot_chains}
\title{Plot MCMC chains for mus and ps}
\usage{
plot_chains(fit, burnin = length(fit$allgens_List)/10)
}
\arguments{
\item{fit}{An object contains all posterior realizations from
\code{\link{est_mix_damcmc}} or \code{\link{est_mix_bdmcmc}}.}

\item{burnin}{Length of burnin, default value is 1/10 of total number of
iteration.}
}
\description{
Plot MCMC chains generated by \code{\link{est_mix_damcmc}}
}
\examples{
# generate data
mix2 <- normmix(ps=c(.4, .6), mus=list(c(0.1, 0.1), c(0.8, 0.8)),
sigmas=list(.02*diag(2), .01*diag(2)))
pp2 <- rsppmix(100,mix2,square(1))
# Run Data augmentation MCMC and get posterior realizations
post=est_mix_damcmc(pp2,L = 5000,2,truncate = F)
# plot MCMC chains for parameter mus and ps
plot_chains(post)
}
\author{
Athanasios Christou Micheas, Jiaxun Chen, Yuchen Wang
}

