% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_contour}
\alias{plot_contour}
\title{Plot contour for intensity surface}
\usage{
plot_contour(mix, pattern, win, L = 100, points = TRUE, filled = TRUE,
  truncate = TRUE, ...)
}
\arguments{
\item{mix}{Object of class normmix.}

\item{pattern}{Object of class \code{\link{ppp}}}

\item{win}{Object of class \code{\link{owin}}}

\item{L}{number of grids on each coordinate.  The default is L=100.}

\item{points}{Logical flag indicating whether plot realizations, i.e.
locations, with the contour plot. The default is TRUE.}

\item{truncate}{Logical flag indicating whether truncation is used for
\code{pattern}. The default is TRUE.}

\item{...}{Further arguments passed to \code{\link[rgl]{filled.contour}}.}
}
\description{
Create a contour plot for the intensity surface with or without realizations
on it.
}
\examples{
# generate a mixture with given ps, mus and sigmas
mix1 <- normmix(ps=c(.5, .5),
                mus=list(c(.2, .2), c(.7, .7)),
                sigmas=list(.01*diag(2), .02*diag(2)))
# generate a realization from mixture mix1
pp1 <- rsppmix(200, mix1, spatstat::square(1))
# plot the theoretical intensity surface with a realization
plot_contour(mix1,pp1,spatstat::square(1))
}

