% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{plot_ind}
\alias{plot_ind}
\title{Plot membership indicator}
\usage{
plot_ind(fit, burnin = length(fit$allgens_List)/10)
}
\arguments{
\item{fit}{An object contains all posterior realizations from
\code{\link{est_mix_damcmc}} or \code{\link{est_mix_bdmcmc}}.}

\item{burnin}{Length of burnin, default value is 1/10 of total number of
iteration.}
}
\description{
Plot membership indicator
}
\examples{
fit <- sppmix::est_mix_damcmc(pp = redwood, m = 3, truncate = FALSE,
                              L = 50000, LL = 100)
plot_ind(fit)
}
\author{
Yuchen Wang, Jiaxun Chen
}

