% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{plot_ind}
\alias{plot_ind}
\title{Plot membership indicator}
\usage{
plot_ind(fit, burnin = fit$L/10)
}
\arguments{
\item{fit}{Object of class \code{damcmc_res} or \code{bdmcmc_res}.}

\item{burnin}{Number of burn-in iterations. By default, it's 1/10 of chain
length.}
}
\description{
Plot membership indicator
}
\examples{
fit <- est_mix_damcmc(pp = redwood, m = 5)
plot_ind(fit)
}

