% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\name{rnormmix}
\alias{rnormmix}
\title{Generate mixture with normal components.}
\usage{
rnormmix(m, sig0, df, rand_m = FALSE, xlim = c(0, 1), ylim = c(0, 1))
}
\arguments{
\item{m}{Number of components in normal mixture.}

\item{sig0}{Tunning parameter in generating random matrix from Wishart
distribution.}

\item{df}{Degree of freedom in generating random matrix from Wishart
distribution.}

\item{rand_m}{whether number of components is random. When
\code{rand_m = FALSE}, it will randomly choose number of components from
\code{1:m}.}

\item{xlim, ylim}{Vector of length two, the limit are used to sample the mu's
from a uniform distribution.}
}
\value{
Object of class \code{normmix}.
}
\description{
Generate a mixture on a 2d window where the mean and variance of the
components are random. The number of component can either be fixed or random.
}
\examples{
mix1 <- rnormmix(m = 3, sig0 = .1, df = 5)
summary(mix1)

mix2 <- rnormmix(m = 5, sig0 = .1, df = 5, rand_m = TRUE, ylim = c(0, 5))
summary(mix2)

}

