% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsppmix.R
\name{rsppmix}
\alias{rsppmix}
\title{Generate a spatial point pattern from normal mixture.}
\usage{
rsppmix(intsurf, truncate = TRUE, lambda, win)
}
\arguments{
\item{truncate}{Whether to truncate the points outside the domain,
default to TRUE.}

\item{lambda}{Intensity of the spatial point pattern}

\item{win}{Object of class spatstat::owin}

\item{mix}{Object of class normmix}
}
\value{
A point pattern of class \code{c("sppmix", "ppp")}.
}
\description{
Generate a spatial point pattern from a given mixture with normal components.
}
\details{
If \code{truncate = FALSE} is set, the returned point pattern will not check
whether the points are inside the domain.

The number of points \code{n} follows Poisson distribution with intensity
lambda * area of window.

When \code{truncate = TRUE}, a point pattern with \code{n} points will be
generated from the mixture first. Then if not all the points are in the
domain, it will generate another \code{n} points until there are more than
 \code{n} points in the domain. The first \code{n} points are returned as
  the generated spatial point pattern.
}
\examples{
# generate a mixture with given ps, mus and sigmas
mix1 <- normmix(ps=c(.5, .5),
                mus=list(c(.2, .2), c(.7, .7)),
                sigmas=list(.01*diag(2), .02*diag(2)))
rsppmix(200, mix1, spatstat::square(1))

# Using different window
rsppmix(200, mix1, spatstat::square(2))

# If truncate = FALSE it will generate points outside the window
rsppmix(200, mix1, spatstat::square(1), truncate = FALSE)
}

