% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppmix.R
\docType{package}
\name{sppmix}
\alias{sppmix}
\alias{sppmix-package}
\title{sppmix: Poisson point process modeling using normal mixture models}
\description{
Modeling spatial point patterns using Poisson point processes with normal
mixtures.
}
\details{
This package implemented classes and methods for modeling spatial
point process data using mixtures of normal component.

Based on \code{spatstat} package's \code{ppp} class for point pattern and
\code{owin} class for window, we implemented \code{normmix} class for
dealing with 2D normal mixture.

Data augmentation MCMC (DAMCMC) and birth-death MCMC (BDMCMC) are the two
main methods we have for estimating normal mixture to point pattern data.

The MCMC algorithms are implemented in C++ using \code{Rcpp} and
\code{RcppArmadillo}, and it's significantly faster than some other
implementations.

\code{rgl} package is used to create 3D intensity plots for intensity surface
from normal mixture.

To learn more about \code{sppmix}, start with those vignettes:
browseVignettes(package = "sppmix")
}

