% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixLabelswitching.R
\name{test_labswitch}
\alias{test_labswitch}
\title{Test if posterior realizations of mu have label switching}
\usage{
test_labswitch(genmus)
}
\arguments{
\item{genmus}{posterior realizations of mus from DAMCMC algorithm.}
}
\description{
Test if there is a label switching of the posterior realizations of mu by
testing if the mean of each component changed dramatically during MCMC.
}
\author{
Athanasios (Sakis) Micheas
}

