% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\name{to_int_surf}
\alias{to_int_surf}
\title{Convert normal mixture to intensity surface.}
\usage{
to_int_surf(mix, lambda = NULL, win = NULL, return_normmix = FALSE)
}
\arguments{
\item{mix}{Object with class normmix or intensity_surface.}

\item{lambda}{Optional parameter of average intensity.}

\item{win}{Optional parameter of domain.}

\item{return_normmix}{Whether to return a normal mixture. (disgard lambda
and win).}
}
\description{
This function can convert a normmix object into an intensity surface. It can
also be used to change intensity or window of an intensity_surface object.
}
\details{
If the class of mix is normmix, lambda and win are used to convert mix to an
intensity surface class. If the class of mix is intensity_surface, lambda
and win are used to change the original setting of lambda and win.
}
\examples{
# from normmix
to_int_surf(demo_mix, lambda = 100, win = square(1))

# from intensity_surface
to_int_surf(demo_intsurf, win = square(2))
to_int_surf(demo_intsurf, lambda = 50)

}

