% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normmix.R
\name{to_int_surf}
\alias{to_int_surf}
\title{Convert normmix object to intensity_surface object or change lambda or window
for intensity_surface object}
\usage{
to_int_surf(mix, lambda = NULL, win = NULL)
}
\arguments{
\item{mix}{object with class normmix or intensity_surface. See Detail and Example.}

\item{lambda}{optional parameter of average intensity.}

\item{win}{optional parameter of domain.}
}
\description{
Convert normmix object to intensity_surface object or change lambda or window
for intensity_surface object
}
\details{
If the class of mix is normmix, lambda and win are used to convert mix to an
intensity surface class. If the class of mix is intensity_surface, lambda
and win are used to change the original setting of lambda and win.
}

